fatal: Split out fatal from User, stub CheckRepairStatus
